Implementando Spring Security e autenticação no projeto

+++ Trabalhos/ProgramacaoParaInternet/pdv-api/src/main/java/org/example/pdvapi/services/ClienteService.java
@@ -3,9 +3,11 @@ package org.example.pdvapi.services;
 import org.example.pdvapi.entity.ClienteEntity;
 import org.example.pdvapi.repositories.ClienteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ClienteService {
@@ -13,30 +15,52 @@ public class ClienteService {
     @Autowired
     private ClienteRepository clienteRepository;
 
-    public ClienteEntity insert(ClienteEntity cliente) {
+    public ResponseEntity<ClienteEntity> insert(ClienteEntity cliente) {
         clienteRepository.save(cliente);
-        return cliente;
+        return ResponseEntity.ok(cliente);
     }
 
-    public ClienteEntity update(ClienteEntity cliente) {
+    public ResponseEntity<ClienteEntity> update(Long id, ClienteEntity cliente) {
+        Optional<ClienteEntity> clienteEntity = clienteRepository.findById(id);
+        if (!clienteEntity.isPresent()) {
+            return ResponseEntity.notFound().build();
+        }
+        cliente.setId(id);
         clienteRepository.save(cliente);
-        return cliente;
+        return ResponseEntity.ok(cliente);
     }
 
-    public void delete(Long id) {
+    public ResponseEntity<Void> delete(Long id) {
+        Optional<ClienteEntity> clienteEntity = clienteRepository.findById(id);
+        if (!clienteEntity.isPresent()) {
+            return ResponseEntity.notFound().build();
+        }
         clienteRepository.deleteById(id);
+        return ResponseEntity.ok().build();
     }
 
-    public ClienteEntity findById(Long id) {
-        return clienteRepository.findById(id).orElse(null);
+    public ResponseEntity<ClienteEntity> findById(Long id) {
+        Optional<ClienteEntity> cliente = clienteRepository.findById(id);
+        if (!cliente.isPresent()) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(cliente.get());
     }
 
-    public List<ClienteEntity> findAll() {
-        return clienteRepository.findAll();
+    public ResponseEntity<List<ClienteEntity>> findAll() {
+        List<ClienteEntity> clientes = clienteRepository.findAll();
+        if (clientes.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(clientes);
     }
 
-    public List<ClienteEntity> findByNomeContaining(String nome) {
-        return clienteRepository.findByNomeContaining(nome);
+    public ResponseEntity<List<ClienteEntity>> findByNomeContaining(String nome) {
+        List<ClienteEntity> clientes = clienteRepository.findByNomeContaining(nome);
+        if (clientes.isEmpty()) {
+            return  ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(clientes);
     }
 
 }
